backend/requirements.txt
aiounittest==1.4.0
click==7.1.2
fastapi==0.63.0
h11==0.12.0
motor==2.3.1
pipdeptree==2.0.0
pydantic==1.8.1
pymongo==3.11.3
python-dotenv==0.15.0
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.4
wrapt==1.12.1
